# Faster SIMD _exp2