@ChrisRackauckas 2021-05-09
This is only used for scalar exponents BTW.